Inverse go to -1 dimensions

+++ src/vector.h
@@ -16,7 +16,7 @@ namespace aline
         // default constructor
         Vector()
         {
-            for (size_t i = 0; i < N; i++)
+            for (int i = 0; i < N; i++)
                 elmts[i] = 0;
         }
 
@@ -42,7 +42,7 @@ namespace aline
         // Copy constructor
         Vector(const Vector<T, N> &v)
         {
-            for (size_t i = 0; i < N; i++)
+            for (int i = 0; i < N; i++)
             {
                 elmts[i] = v.elmts[i];
             }
